script: Extract target_tx.go file
This commit extracts the TargetTX implementation into its own file. There are
no functional changes in this commit.
